Add a method to get the operands of all instructions
This commit introduces a method that returns the operand types of all
instructions.  Each element of the array corresponds to an instruction
with the same index in the `RubyVM::INSTRUCTIONS_NAMES` array

For example, the instruction at offset 5 in `RubyVM::INSTRUCTIONS_NAMES`
will have the operands at the same offset in the array returned by the
`RubyVM.instruction_operands` list.