Free up two bits on callcache imemos
We need 16 bits in the header to store shape ids in the call cache on
32bit machines.  rb_callcache objects were using two bits for detecting
if the object was a global or allocated on the stack.  This commit moves
those two bits to the "frozen" flag as well as the "exivar" flag.  I
think this is OK because these callcache objects should never be mutated
(since they are stack / globally allocated), and they should never be
seen by the GC (the flags indicate that we should never pass these
pointers to the GC)