Fix MJIT + Shapes
This commit exposes the shape bits as a constant on the `RubyVM::Shape`
class.  The number of bits the shape consumes depends on the platform as
well as whether or not RUBY_DEBUG is enabled.  If RUBY_DEBUG is enabled
we define the shape id as an uint32_t, but we can only consume the upper
16 bits because Ractors use the lower 16 bits for verification.
Exposing the shape id as a constant in Ruby means we don't need to
expose the value of RUBY_DEBUG to anyone