Remove useless checks
We don't need to check the number of IVs.  If the shape was set on the
object, it necessarily must have the room to read the IV (the room was
guaranteed to exist at the time the shape was set on the object).
Additionally, we don't need to test EXIVAR because we were able to look
up the IV index (the IV index existing guarantees the IV has been set)