WIP: Transition shapes on capacity changes
We transition shapes on capacity changes, and set the cache to
invalid when capacity changes.

This allows us to remove the check for capacity from the fastpath,
only growing an object's IV list in the slow path.

Still todo: Initializing the size_pool_shape_ids is not happening
in exactly the right place.