Set ivars individually on dups allocated out of different size_pools
When copying an object, as an optimization, if the objects are allocated
out of the same size pools, we copy the ivar memory and directly copy the
shape ID.

However, if the new object is allocated out of a different size pool, we
cannot do this, because the shapes will be different. In this commit, we
set ivars on the new object one by one to ensure it makes its own shape
transitions.